Fallback to original theme or color_scheme if one of them is missing
As skin validation is changed to only require one of `color_scheme` or `theme` key to be valid, the invalid one of both needs to be set to a defined value. Otherwise the most recent highlighted theme or color scheme will be commited, if a skin with one of them missing is selected.